pythonfuncs/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
blinker==1.6.2
CacheControl==0.13.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
ci-info==0.3.0
click==8.1.7
cloudevents==1.9.0
configobj==5.0.8
configparser==6.0.0
cryptography==41.0.4
deprecation==2.1.0
dnspython==2.4.2
etelemetry==0.3.0
filelock==3.12.4
firebase-admin==6.2.0
firebase-functions==0.1.1
fits==0.0.1
Flask==2.3.3
Flask-Cors==4.0.0
frozenlist==1.4.0
functions-framework==3.4.0
functions-wrapper==1.0.2
future==0.18.3
google-api-core==2.12.0
google-api-python-client==2.102.0
google-auth==2.23.2
google-auth-httplib2==0.1.1
google-cloud-core==2.3.3
google-cloud-firestore==2.12.0
google-cloud-storage==2.11.0
google-crc32c==1.5.0
google-events==0.11.0
google-resumable-media==2.6.0
googleapis-common-protos==1.60.0
grpcio==1.59.0
grpcio-status==1.59.0
gunicorn==20.1.0
httplib2==0.22.0
idna==3.4
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
loguru==0.7.2
looseversion==1.3.0
lxml==4.9.3
MarkupSafe==2.1.3
msgpack==1.0.7
multidict==6.0.4
networkx==3.1
nibabel==5.1.0
nipype==1.8.6
numpy==1.26.0
openai==0.28.1
packaging==23.2
pandas==2.1.1
pathlib==1.0.1
pinecone-client==2.2.4
proto-plus==1.22.3
protobuf==4.24.4
prov==2.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydot==1.4.2
PyJWT==2.8.0
PyMuPDF==1.23.4
PyMuPDFb==1.23.3
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pyxnat==1.6
PyYAML==6.0.1
rdflib==7.0.0
requests==2.31.0
rsa==4.9
scipy==1.11.3
simplejson==3.19.1
six==1.16.0
tqdm==4.66.1
traits==6.3.2
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.6
watchdog==3.0.0
Werkzeug==3.0.0
yarl==1.9.2